Fix (?) zipping of symlinked directory
It turns out that zipping a relative path will preserve that path in the zip archive, so zipping e.g. ../compiler/reader will actually include a "../compiler" directory in the archive.  Also, it appears that zip follows symlinks by default, and the flag --symlinks actually causes it *not* to follow symlinks.